requirements.txt
fastapi
uvicorn[standard]

sqlalchemy
pymysql

python-jose[cryptography]
passlib[bcrypt]

pydantic
pydantic-settings

kafka-python
python-multipart
